requirements.txt
# LangChain 1.x stack (required for langchain-google-genai 4.x; 0.2.x conflicts)
# See: https://docs.langchain.com/oss/python/integrations/chat/google_generative_ai
langchain>=1.0,<2
langchain-community>=0.4
langchain-core>=1.2.5
# langchain-openai  # commented out; using Gemini
# openai
langchain-google-genai>=4.0.0
httpx>=0.27.0
requests>=2.32.5
python-dotenv==1.0.1
rank-bm25>=0.2.2
mcp
